Configuration

+++ test/advent/cli_test.rb
@@ -26,8 +26,29 @@ class Advent::CLITest < Advent::TestCase
       "day 5 input"
     )
 
-    @cli = Advent::CLI.new([], root_path: DUMMY_ROOT_PATH, http_module: http_mock)
-    @year_cli = Advent::CLI.new([], root_path: DUMMY_ROOT_PATH.join("2015"))
+    Dir.chdir DUMMY_ROOT_PATH
+    @cli = Advent::CLI.new([], http_module: http_mock)
+  end
+
+  def teardown
+    Dir.chdir DUMMY_ROOT_PATH
+    Advent.session.clear
+  end
+
+  def test_init
+    path = DUMMY_ROOT_PATH.join("init")
+    Dir.mkdir path
+
+    out, _err = capture_io do
+      Dir.chdir path do
+        @cli.invoke(:init)
+      end
+    end
+
+    assert_match(/create.*advent.yml/, out.strip)
+  ensure
+    File.delete path.join("advent.yml")
+    Dir.rmdir path
   end
 
   def test_version
@@ -38,7 +59,7 @@ class Advent::CLITest < Advent::TestCase
     assert_equal Advent::VERSION, out.strip
   end
 
-  def test_solve_from_parent_directory
+  def test_solve_from_root_directory
     out, _err = capture_io do
       @cli.invoke(:solve, ["2015/day1.rb"])
     end
@@ -48,13 +69,15 @@ class Advent::CLITest < Advent::TestCase
 
   def test_solve_from_year_directory
     out, _err = capture_io do
-      @year_cli.invoke(:solve, ["day2.rb"])
+      Dir.chdir DUMMY_ROOT_PATH.join("2015") do
+        @cli.invoke(:solve, ["day2.rb"])
+      end
     end
 
     assert_equal "Part 1: 789\nPart 2: Missing", out.strip
   end
 
-  def test_generate_solution_with_year_and_day
+  def test_generate
     capture_io do
       @cli.invoke(:generate, ["2015", "3"])
     end
@@ -67,9 +90,26 @@ class Advent::CLITest < Advent::TestCase
     end
   end
 
-  def test_generate_solution_from_year_directory_with_day
+  def test_download_when_generating
+    with_session("abc123") do
+      with_config({"download_when_generating" => true}) do
+        capture_io do
+          @cli.invoke(:generate, ["2015", "3"])
+        end
+      end
+    end
+
+    assert File.exist? DUMMY_ROOT_PATH.join("2015", "day3.rb")
+    assert File.exist? DUMMY_ROOT_PATH.join("2015", ".day3.input.txt")
+
+    File.delete DUMMY_ROOT_PATH.join("2015", "day3.rb")
+    File.delete DUMMY_ROOT_PATH.join("2015", "test", "day3_test.rb")
+    File.delete DUMMY_ROOT_PATH.join("2015", ".day3.input.txt")
+  end
+
+  def test_generate_day_parsing
     capture_io do
-      @year_cli.invoke(:generate, ["3"])
+      @cli.invoke(:generate, ["2015", "day3"])
     end
 
     ["day3.rb", "test/day3_test.rb"].each do |file_name|
@@ -80,20 +120,7 @@ class Advent::CLITest < Advent::TestCase
     end
   end
 
-  def test_generate_solution_with_non_numbers_in_day
-    capture_io do
-      @year_cli.invoke(:generate, ["day3"])
-    end
-
-    ["day3.rb", "test/day3_test.rb"].each do |file_name|
-      expected_file_path = DUMMY_ROOT_PATH.join("2015", file_name)
-      assert File.exist? expected_file_path
-
-      File.delete expected_file_path
-    end
-  end
-
-  def test_generate_solution_valid_minimum_year
+  def test_generate_valid_minimum_year
     _out, err = capture_io do
       @cli.invoke(:generate, ["2013", "1"])
     end
@@ -101,15 +128,15 @@ class Advent::CLITest < Advent::TestCase
     assert_equal "Advent of Code only started in 2014!", err.strip
   end
 
-  def test_generate_solution_valid_maximum_year
+  def test_generate_valid_maximum_year
     _out, err = capture_io do
-      @cli.invoke(:generate, [Date.today.year + 1, "1"])
+      @cli.invoke(:generate, [(Date.today.year + 1).to_s, "1"])
     end
 
     assert_equal "Future years are not supported.", err.strip
   end
 
-  def test_generate_solution_valid_minimum_day
+  def test_generate_valid_minimum_day
     _out, err = capture_io do
       @cli.invoke(:generate, ["2015", "0"])
     end
@@ -117,7 +144,7 @@ class Advent::CLITest < Advent::TestCase
     assert_equal "Day must be between 1 and 25 (inclusive).", err.strip
   end
 
-  def test_generate_solution_valid_maximum_day
+  def test_generate_valid_maximum_day
     _out, err = capture_io do
       @cli.invoke(:generate, ["2015", "26"])
     end
@@ -125,9 +152,9 @@ class Advent::CLITest < Advent::TestCase
     assert_equal "Day must be between 1 and 25 (inclusive).", err.strip
   end
 
-  def test_download_input
+  def test_download_from_root_directory
     out, _err = capture_io do
-      with_stdin_input(@session) do
+      with_session(@session) do
         @cli.invoke(:download, ["2015", "3"])
       end
     end
@@ -137,10 +164,10 @@ class Advent::CLITest < Advent::TestCase
     assert_equal "day 3 input", File.read(input_path)
     assert_match(/Input downloaded to #{input_path}/, out.strip)
   ensure
-    File.delete input_path
+    File.delete input_path if File.exist? input_path
   end
 
-  def test_download_persists_session_cookie
+  def test_persisting_session_cookie
     with_stdin_input(@session) do
       capture_io do
         @cli.invoke(:download, ["2015", "4"])
@@ -149,11 +176,11 @@ class Advent::CLITest < Advent::TestCase
 
     assert_equal @session, Advent.session.value
   ensure
-    File.delete DUMMY_ROOT_PATH.join("2015", ".day4.input.txt")
-    Advent.session.clear
+    input = DUMMY_ROOT_PATH.join("2015", ".day4.input.txt")
+    File.delete input if File.exist? input
   end
 
-  def test_download_skips_asking_for_session_cookie
+  def test_not_asking_for_session_cookie_again
     Advent.session.value = @session
 
     out, _err = capture_io do
@@ -164,22 +191,20 @@ class Advent::CLITest < Advent::TestCase
 
     refute_match(/session cookie value/, out)
   ensure
-    File.delete DUMMY_ROOT_PATH.join("2015", ".day5.input.txt")
-    Advent.session.clear
+    input = DUMMY_ROOT_PATH.join("2015", ".day5.input.txt")
+    File.delete input if File.exist? input
   end
 
-  def test_download_input_failure
+  def test_download_error
     _out, err = capture_io do
-      with_stdin_input("invalid") do
+      with_session("invalid") do
         @cli.invoke(:download, ["2015", "6"])
       end
     end
 
-    input_path = DUMMY_ROOT_PATH.join("2015", ".day6.input.txt")
-    refute File.exist?(input_path)
     assert_match(/Something went wrong/, err.strip)
   ensure
-    File.delete input_path if File.exist? input_path
-    Advent.session.clear
+    input = DUMMY_ROOT_PATH.join("2015", ".day6.input.txt")
+    File.delete input_path if File.exist? input
   end
 end
